Refactored and put helper functions in header files
